+++ bot/src/main/java/edu/java/bot/domain/links/ViewLinksUseCase.java
@@ -2,7 +2,7 @@ package edu.java.bot.domain.links;
 
 import com.pengrad.telegrambot.model.User;
 import edu.java.bot.data.LinkTrackerRepository;
-import edu.java.bot.model.UserIsNotAuthenticated;
+import edu.java.core.exception.UserIsNotAuthenticated;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
