--- bot/src/main/java/edu/java/bot/model/UserAlreadyRegisteredException.java
@@ -1,3 +1,0 @@
-package edu.java.bot.model;
-
-public class UserAlreadyRegisteredException extends RuntimeException {}
