+++ bot/src/main/java/edu/java/bot/view/controller/impl/LinkControllerImpl.java
@@ -1,6 +1,6 @@
 package edu.java.bot.view.controller.impl;
 
-import edu.java.bot.model.dto.LinkUpdateRequest;
+import edu.java.core.request.LinkUpdateRequest;
 import edu.java.bot.view.controller.LinkController;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
