+++ core/src/main/java/edu/java/core/exception/LinkAlreadyNotTracked.java
@@ -1,3 +1,3 @@
-package edu.java.bot.model;
+package edu.java.core.exception;
 
 public class LinkAlreadyNotTracked extends RuntimeException {}
