+++ core/src/main/java/edu/java/core/exception/LinkAlreadyTracked.java
@@ -1,3 +1,3 @@
-package edu.java.bot.model;
+package edu.java.core.exception;
 
 public class LinkAlreadyTracked extends RuntimeException {}
