+++ core/src/main/java/edu/java/core/exception/UserAlreadyAuthenticated.java
@@ -1,0 +1,3 @@
+package edu.java.core.exception;
+
+public class UserAlreadyAuthenticated extends RuntimeException {}
