+++ core/src/main/java/edu/java/core/request/LinkUpdateRequest.java
@@ -1,4 +1,4 @@
-package edu.java.bot.model.dto;
+package edu.java.core.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
