+++ core/src/main/java/edu/java/core/response/ListLinksResponse.java
@@ -1,17 +1,13 @@
-package edu.java.bot.model.dto;
+package edu.java.core.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.NonNull;
-
 import java.util.List;
 
 public record ListLinksResponse(
     @JsonProperty("links")
-    @NonNull
     List<LinkResponse> links,
 
     @JsonProperty("size")
-    @NonNull
     Integer size
 ) {
 }
