+++ core/src/main/java/edu/java/core/response/github/GithubRepositoryResponse.java
@@ -1,4 +1,4 @@
-package edu.java.scrapper.model.github;
+package edu.java.core.response.github;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
