+++ scrapper/src/main/java/edu/java/scrapper/data/client/GithubClient.java
@@ -1,6 +1,6 @@
 package edu.java.scrapper.data.client;
 
-import edu.java.scrapper.model.github.GithubRepositoryResponse;
+import edu.java.core.response.github.GithubRepositoryResponse;
 
 public interface GithubClient {
     GithubRepositoryResponse fetchRepository(String user, String repository);
