+++ scrapper/src/main/java/edu/java/scrapper/data/client/StackOverflowClient.java
@@ -1,6 +1,6 @@
 package edu.java.scrapper.data.client;
 
-import edu.java.scrapper.model.stackoverflow.StackOverflowAnswersResponse;
+import edu.java.core.response.stackoverflow.StackOverflowAnswersResponse;
 
 public interface StackOverflowClient {
     StackOverflowAnswersResponse fetchAnswers(String questionId);
