+++ scrapper/src/main/java/edu/java/scrapper/data/client/impl/StackOverflowClientImpl.java
@@ -1,7 +1,7 @@
 package edu.java.scrapper.data.client.impl;
 
 import edu.java.scrapper.data.client.StackOverflowClient;
-import edu.java.scrapper.model.stackoverflow.StackOverflowAnswersResponse;
+import edu.java.core.response.stackoverflow.StackOverflowAnswersResponse;
 import edu.java.scrapper.di.util.ApiQualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
