--- scrapper/src/main/java/edu/java/scrapper/model/LinkResponse.java
@@ -1,11 +1,0 @@
-package edu.java.scrapper.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public record LinkResponse(
-    @JsonProperty("id")
-    Long id,
-    @JsonProperty("url")
-    String url
-) {
-}
