--- scrapper/src/main/java/edu/java/scrapper/model/ListLinksResponse.java
@@ -1,13 +1,0 @@
-package edu.java.scrapper.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
-
-public record ListLinksResponse(
-    @JsonProperty("links")
-    List<LinkResponse> links,
-
-    @JsonProperty("size")
-    Integer size
-) {
-}
